Cross Cutting - Service Registry Discovery Pattern Cross Cutting - API Gateway Pattern e Load Balance Transação Distribuída para Deleção de Recursos entre Microservices

+++ src/main/java/com/ead/course/services/impl/CourseServiceImpl.java
@@ -1,8 +1,10 @@
 package com.ead.course.services.impl;
 
 import com.ead.course.models.CourseModel;
+import com.ead.course.models.CourseUserModel;
 import com.ead.course.models.ModuleModel;
 import com.ead.course.repositories.CourseRepository;
+import com.ead.course.repositories.CourseUserRepository;
 import com.ead.course.repositories.LessonRepository;
 import com.ead.course.repositories.ModuleRepository;
 import com.ead.course.services.CourseService;
@@ -16,7 +18,6 @@ import javax.transaction.Transactional;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
-import java.util.stream.Collectors;
 
 @Service
 public class CourseServiceImpl implements CourseService {
@@ -30,31 +31,29 @@ public class CourseServiceImpl implements CourseService {
     @Autowired
     LessonRepository lessonRepository;
 
+    @Autowired
+    CourseUserRepository courseUserRepository;
+
     @Transactional
     @Override
     public void delete(CourseModel courseModel) {
-        List<ModuleModel> modules = moduleRepository.findAllModulesIntoCourse(courseModel.getCourseId())
-                .stream()
-                .collect(Collectors.toList());
+        List<ModuleModel> modules = moduleRepository.findAllModulesIntoCourse(courseModel.getCourseId());
 
         modules.forEach(lesson ->
             lessonRepository.deleteAll(lessonRepository.findAllLessonsIntoModule(lesson.getModuleId()))
         );
 
-        moduleRepository.deleteAll(modules);
-        courseRepository.delete(courseModel);
+        List<CourseUserModel> courseUserModelList = courseUserRepository.findAllCourseUserIntoCourse(courseModel.getCourseId());
 
-//        List<ModuleModel> moduleModelList = moduleRepository.findAllLModulesIntoCourse(courseModel.getCourseId());
-//        if (!moduleModelList.isEmpty()){
-//            for(ModuleModel module : moduleModelList){
-//                List<LessonModel> lessonModelList = lessonRepository.findAllLessonsIntoModule(module.getModuleId());
-//                if (!lessonModelList.isEmpty()){
-//                    lessonRepository.deleteAll(lessonModelList);
-//                }
-//            }
-//            moduleRepository.deleteAll(moduleModelList);
-//        }
-//        courseRepository.delete(courseModel);
+        if(!modules.isEmpty()){
+            moduleRepository.deleteAll(modules);
+        }
+
+        if(!courseUserModelList.isEmpty()){
+            courseUserRepository.deleteAll(courseUserModelList);
+        }
+
+        courseRepository.delete(courseModel);
     }
 
     @Override
@@ -71,6 +70,5 @@ public class CourseServiceImpl implements CourseService {
     public Page<CourseModel> findAll(Specification<CourseModel> spec, Pageable pageable) {
         return courseRepository.findAll(spec, pageable);
     }
-
-
+    
 }
